test(reorder): recategorize and clean up browser tests
